refactor: files into a different dir
